requirements.txt
fastapi
uvicorn
pydantic
python-multipart
mysql-connector-python
python-dotenv
httpx
